Preferences moved to JVM main. Compiles, but with proto issues.
BundleInsideHelper is not working with kmp project so all proto related things are commented out with failing tests.
