+++ datastore/datastore-preferences-core/src/jvmTest/kotlin/androidx/datastore/preferences/core/PreferencesFromJavaTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Android Open Source Project
+ * Copyright 2022 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -80,7 +80,7 @@ public class PreferencesFromJavaTest {
         MutablePreferences mutablePreferences = PreferencesFactory.createMutable();
         Preferences.Key<Integer> integerKey = PreferencesKeys.intKey("integer_key");
 
-        assertNull(mutablePreferences.get(integerKey));
+        Assert.assertNull(mutablePreferences.get(integerKey));
     }
 
     @Test
